waitm-637: restructure story
